python/requirements.txt
python-dotenv
openai
requests
websockets
twitchio
